refactor: store literal byte offset in hash table

+++ hashtable.c
@@ -1,6 +1,8 @@
+#include <stdio.h>
+
 struct HashEntry {
     char path[4096];
-    int offset;
+    off_t offset;
 };
 
 struct HashEntry **hash_table;
